+++ internal/dnsforward/dnsforward.go
@@ -56,6 +56,10 @@ type Server struct {
 	stats      stats.Stats
 	access     *accessCtx
 
+	// autohostSuffix is the suffix used to detect internal hosts.  It must
+	// be a valid top-level domain plus dots on each side.
+	autohostSuffix string
+
 	ipset ipsetCtx
 
 	tableHostToIP     map[string]net.IP // "hostname -> IP" table for internal addresses (DHCP)
@@ -74,21 +78,50 @@ type Server struct {
 	conf ServerConfig
 }
 
-// DNSCreateParams - parameters for NewServer()
+// defaultAutohostSuffix is the default suffix used to detect internal hosts
+// when no suffix is provided.  See the documentation for Server.autohostSuffix.
+const defaultAutohostSuffix = ".lan."
+
+// DNSCreateParams are parameters to create a new server.
 type DNSCreateParams struct {
-	DNSFilter  *dnsfilter.DNSFilter
-	Stats      stats.Stats
-	QueryLog   querylog.QueryLog
-	DHCPServer dhcpd.ServerInterface
+	DNSFilter   *dnsfilter.DNSFilter
+	Stats       stats.Stats
+	QueryLog    querylog.QueryLog
+	DHCPServer  dhcpd.ServerInterface
+	AutohostTLD string
+}
+
+// tldToSuffix converts a top-level domain into an autohost suffix.
+func tldToSuffix(tld string) (suffix string) {
+	l := len(tld) + 2
+	b := make([]byte, l)
+	b[0] = '.'
+	copy(b[1:], tld)
+	b[l-1] = '.'
+
+	return string(b)
 }
 
 // NewServer creates a new instance of the dnsforward.Server
 // Note: this function must be called only once
-func NewServer(p DNSCreateParams) *Server {
-	s := &Server{
-		dnsFilter: p.DNSFilter,
-		stats:     p.Stats,
-		queryLog:  p.QueryLog,
+func NewServer(p DNSCreateParams) (s *Server, err error) {
+	var autohostSuffix string
+	if p.AutohostTLD == "" {
+		autohostSuffix = defaultAutohostSuffix
+	} else {
+		err = validateDomainNameLabel(p.AutohostTLD)
+		if err != nil {
+			return nil, fmt.Errorf("autohost tld: %w", err)
+		}
+
+		autohostSuffix = tldToSuffix(p.AutohostTLD)
+	}
+
+	s = &Server{
+		dnsFilter:      p.DNSFilter,
+		stats:          p.Stats,
+		queryLog:       p.QueryLog,
+		autohostSuffix: autohostSuffix,
 	}
 
 	if p.DHCPServer != nil {
@@ -101,7 +134,8 @@ func NewServer(p DNSCreateParams) *Server {
 		// Use plain DNS on MIPS, encryption is too slow
 		defaultDNS = defaultBootstrap
 	}
-	return s
+
+	return s, nil
 }
 
 // NewCustomServer creates a new instance of *Server with custom internal proxy.