+++ internal/home/whois_test.go
@@ -6,15 +6,23 @@ import (
 
 	"github.com/AdguardTeam/AdGuardHome/internal/dnsforward"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
-func prepareTestDNSServer() error {
+func prepareTestDNSServer(t *testing.T) {
+	t.Helper()
+
 	config.DNS.Port = 1234
-	Context.dnsServer = dnsforward.NewServer(dnsforward.DNSCreateParams{})
+
+	var err error
+	Context.dnsServer, err = dnsforward.NewServer(dnsforward.DNSCreateParams{})
+	require.NoError(t, err)
+
 	conf := &dnsforward.ServerConfig{}
 	conf.UpstreamDNS = []string{"8.8.8.8"}
 
-	return Context.dnsServer.Prepare(conf)
+	err = Context.dnsServer.Prepare(conf)
+	require.NoError(t, err)
 }
 
 // TODO(e.burkov): It's kind of complicated to get rid of network access in this
@@ -22,12 +30,15 @@ func prepareTestDNSServer() error {
 // the server, so it becomes hard to simulate handling of request from test even
 // with substituted upstream.  However, it must be done.
 func TestWhois(t *testing.T) {
-	assert.Nil(t, prepareTestDNSServer())
+	prepareTestDNSServer(t)
 
 	w := Whois{timeoutMsec: 5000}
 	resp, err := w.queryAll(context.Background(), "8.8.8.8")
-	assert.Nil(t, err)
+	assert.NoError(t, err)
+
 	m := whoisParse(resp)
+	require.NotEmpty(t, m)
+
 	assert.Equal(t, "Google LLC", m["orgname"])
 	assert.Equal(t, "US", m["country"])
 	assert.Equal(t, "Mountain View", m["city"])